Implement a TD_Data_Dictionary data structure
This commit implements a wrapper data structure I am dubbing a
TD_Data_Dictionary. Put simply this data structure is a dictionary whose
keys are TD_Data_Keys and whose values are TD_Data_Values.

A TD_Data_Key is an object with a string of the key that is pressed or
released
A TD_Data_Value is an object containing a list guaranteed to have two
critical elements:
    1) The action performed on the key (Press or release)
    2) The time that the action was performed

This data structure was motivated to make storing and interacting with
the typing dynamics data from the csv file easier and cleaner, rather
than interacting with a pandas dataframe.

In that spirit, this commit also implements a new version of the
calculate_key_hit_time() function and lays the foundations for the
calculate_key_interval_time() function. The implementation of
calculate_key_hit_time() is still a little messy and complicated but its
diff compared to the pandas version has saved us ~30 ish lines, if you
don't include the explanatory comment in the middle of the code, so I
think this change was well worth it