Implement a basic evaluator for the verifiers.
This commit implements a basic evaluator for the verifiers by simply
checking the number of valid keys that they found for both the KHT and
KIT feature sets versus the total number of potential matching keys
that both feature sets. It then computes the quotient and sees if this
value exceeds a threshold. If it does then it is considered okay.

There is probably more we can do to imporve the robustness of the
evaluator but this should be a good starting point