Add basic html reporting and generation support
This commit allows support to generate pretty html reports and tables
for data and information generated from this project. We do this using
the Datapane library

This could be useful if, for example, we wanted to give a table of all
the data  present in two typing samples and then show in a separate
table the matching entries we found. We could also figure out ways to
creatively display information generated from the verifiers, in theory.