Rework pickle_driver API for easier file reading
This commit reworks the API in PickleDriver to make file reading
simpler. However, the cost is we have to generate more files, but I
think that cost should be somewhat amortized with how small pickled data
is
The new API seperates out writing KIT and KHT data to files into
seperate functions. The functions will also internally modify the names
of the files generated at the provided path argument to make clear what
data is held in the file (KHT_ or KIT_). The API will also ammed what
form of KIT (PP, PR, RP, RR) to the filename as well. The provided
read function luckily does not need to change. However, we should
probably  make it a little easier for the user to know the name/filepath
of the generated file