AbsoluteVerifier can run against KHT (and KIT a little bit) from files
This commit rearchitects the AbsoluteVerifier to support running against
files containing KHT and KIT data rather than computing them on the fly.

The KHT on first inspection seems to work as intended, however, when
trying with KIT there is drastic differences between accepted values. So
I don't know if i broke the original way without evaluation or the new
way so we need ti test it more